template/base: copy scripts after installs
A majority of the time while updating an image, the shell scripts baked
in the image are what are being changed. Having this stage before the
downloads forces all packages to be downloaded everytime you make a
change to the script.

Moving the copying of these scripts after the downloads will help
developers test their code changes faster.