template/base: prevent race condition between jobs
According to the Kubernetes documentation [1], the `metadata.resourceVersion`
field in Kubernetes resources can be used as a locking mechanism to
prevent multiple applies from overwriting each other.

This is achieved by providing `resourceVersion` in your k8s manifest,
and it does not match the version stored in etcd, the apply will fail
with a StatusConflict (HTTP status code 409).

This change should prevent race conditions between jobs running at the
same time, which should also prevent resources being unexpectedly
deleted because of this race condition and the way the resource deletion
logic is implemented.

[1] https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency